+++ element/image/classes/element.php
@@ -31,7 +31,6 @@ use context_system;
 use core_collator;
 use html_writer;
 use mod_customcert\service\form_service;
-use mod_customcert\element\constructable_element_interface;
 use mod_customcert\element\persistable_element_interface;
 use mod_customcert\element as base_element;
 use mod_customcert\element\renderable_element_interface;
@@ -57,7 +56,6 @@ use stored_file;
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 class element extends base_element implements
-    constructable_element_interface,
     copyable_element_interface,
     form_element_interface,
     persistable_element_interface,
@@ -88,15 +86,6 @@ class element extends base_element implements
         parent::__construct($element);
     }
 
-    /**
-     * Build an element instance from a DB record.
-     *
-     * @param stdClass $record Raw DB row from customcert_elements.
-     * @return static
-     */
-    public static function from_record(stdClass $record): static {
-        return new static($record);
-    }
 
     /**
      * Build the configuration form for this element.